improve project structure to scale
