+++ server/src/resolvers/stripe/credit-card/index.ts
@@ -1,0 +1,1 @@
+export { change } from "./change";
