+++ server/src/resolvers/stripe/index.ts
@@ -1,0 +1,4 @@
+import * as subscription from "./subscription";
+import * as creditCard from "./credit-card";
+
+export { subscription, creditCard };
