+++ server/src/resolvers/stripe/user.ts
@@ -1,0 +1,1 @@
+export { User } from "../User";
