+++ server/src/resolvers/user.ts
@@ -1,0 +1,1 @@
+export { User } from "../entity/User";
